+++ src/imageprocess/kalman.h
@@ -3,17 +3,17 @@
 #include <opencv2/video/tracking.hpp>
 class KFTracking {
 	cv::KalmanFilter KF;
-	float ts;
-	void set_dt(float dt);
+	double ts;
+	void set_dt(double dt);
 	bool is_init = false;
 public:
-	float dt = 0.05;
+	double dt = 0.05;
 	KFTracking();
-	void init(const cv::Rect2i& face, float ts, float m_noise);
+	void init(const cv::Rect2i& face, double ts, float m_noise);
 
-	void update(const cv::Rect2i& face, float ts, float m_noise);
+	void update(const cv::Rect2i& face, double ts, float m_noise);
 	void reset() {
 		is_init = false; 
 	}
-	cv::Rect2i predict(float dt = 0.05);
+	cv::Rect2i predict(double dt = 0.05);
 };
